requirements/base.txt
# Wheel
wheel>=0.40.0, <0.41

# Setuptools
setuptools>=67.6.0, <67.7

# Packaging
packaging>=23.0, <24.0

# Django
django>=3.2.23, <3.3

# Configuration
django-environ>=0.10.0, <0.11

# Forms
django-crispy-forms>=2.0, <2.1
crispy-bootstrap4==2022.1

# Models
django-model-utils>=4.3.1, <4.4

# Password storage
argon2-cffi>=21.3.0, <21.4

# Python-PostgreSQL Database Adapter
psycopg2-binary>=2.9.5, <2.10

# Unicode slugification
awesome-slugify>=1.6.5, <1.7

# Time zones support
pytz>=2022.7.1

# SVG icon support
django-iconify==0.1.1  # NOTE: v0.3 crashes, see issue

# Online documentation via django-docs
docutils==0.18.1  # NOTE: sphinx-rtd-theme 1.2 requires <0.19
Sphinx==6.2.1  # NOTE: sphinx-rtd-theme v1.2.2 forces <7
django-docs==0.3.3
sphinx-rtd-theme==1.2.2

# Versioning
versioneer==0.28

######################
# Project app imports
######################

# Django-plugins (with Django v3.0+ support)
django-plugins-bihealth==0.4.0

# Rules for permissions
rules>=3.3, <3.4

# REST framework
djangorestframework>=3.14.0, <3.15

# Keyed list addon for DRF
drf-keyed-list-bihealth==0.1.1

# Token authentication
django-rest-knox>=4.2.0, <4.3

# Markdown field support
markdown==3.4.1
django-markupfield>=2.0.1, <2.1
django-pagedown>=2.2.1, <2.3
mistune>=2.0.5, <2.1

# Database file storage for filesfolders
django-db-file-storage==0.5.5

# Celery dependency
redis>=4.4.4

# Backround Jobs requirements
celery>=5.2.7, <5.3

# Django autocomplete light (DAL)
# NOTE: 3.9.5 causes crash with Whitenoise (see issue #1224)
django-autocomplete-light==3.9.4

# SAML2 support for SSO
django-saml2-auth-ai>=2.1.6, <2.2
